AimAtPoseWhilePathing added
need to do one for when robot is stationary
